requirements.txt
flask
qbittorrent-api
python-decouple
tpblite
lxml